Breaking up client god-object into multiple pieces... part 1 of N
This change is the first in the series of refactoring the client object into a set of smaller managable pieces, the guidelines for those changes are:

  - Completely get rid if the `traits` module and organize the traits defined their by their function rather than the type of language construct they are.
  - There are around 10+ different types called `BlockChain` encountered so far during this refactoring session. Reduce them to one.
  - Completely get rid of the `Client` struct and split its functionality into componentized types that interact with each other - preferably through message passing in an agent architecture.

In this first stab of splitting the `client.rs` file into smaller files I have made some internal member variables public so they are accessible from within other files. This will change with time, it's a temporary remedy.